Implementacion: Dettales y correcion, finalizacion

+++ src/app/lista-empleados/lista-empleados.component.ts
@@ -3,6 +3,8 @@ import { NgFor } from '@angular/common';
 import { Empleado } from '../empleado';
 import { EmpleadoService } from '../empleado.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-lista-empleados',
@@ -25,16 +27,41 @@ export class ListaEmpleadosComponent implements OnInit {
     this.router.navigate(['actualizar-empleado', id]); // Redirige a la página de edición
   }
 
-  eliminarEmpleado(id: number) {
-    this.empleadoServicio.eliminarEmpleado(id).subscribe(dato => {
-      console.log(dato);
-      this.obtenerEmpleados(); // Vuelve a cargar la lista después de eliminar
-    })
-  }
-
   private obtenerEmpleados(){
     this.empleadoServicio.ObtenerListaDeEmpleados().subscribe(dato => {
       this.empleados = dato;
     })
+  }
+
+  eliminarEmpleado(id: number) {
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "Confirma si deseas eliminar al empleado",
+      icon: 'warning', // Cambiado 'type' a 'icon'
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, elimínalo',
+      cancelButtonText: 'No, cancelar',
+      buttonsStyling: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.empleadoServicio.eliminarEmpleado(id).subscribe(dato => {
+          console.log(dato);
+          this.obtenerEmpleados();
+          Swal.fire(
+            'Empleado eliminado',
+            'El empleado ha sido eliminado con exito',
+            'success'
+          )
+        })
+      }
+    });
+
+
+  }
+
+  verDetallesDelEmpleado(id: number){
+    this.router.navigate(['empleado-detalles', id]);
   }
 }
